Export missing types from react client
