+++ packages/react-client/src/hooks/internal/useFishjamClientState.ts
@@ -1,7 +1,7 @@
 import type { Component, FishjamClient, GenericMetadata, MessageEvents, Peer } from "@fishjam-cloud/ts-client";
 import { useCallback, useMemo, useRef, useSyncExternalStore } from "react";
 
-import type { PeerId } from "../../types/internal";
+import type { PeerId } from "../../types/public";
 
 const eventNames = [
   "socketClose",
